+++ t/004_signatures.c
@@ -18,6 +18,7 @@
 
 #define DBLTAP_IMPLEMENTATION
 #include "common/double_tap.h"
+#include <ctype.h>
 #include <infix/infix.h>
 #include <string.h>
 
@@ -56,8 +57,57 @@ static void test_type_fail(const char * signature, const char * name) {
 
 void dummy_handler() {}
 
+// Helper to remove all whitespace from a string for canonical comparison.
+static void normalize_string(char * s) {
+    if (!s)
+        return;
+    char * d = s;
+    do {
+        while (isspace((unsigned char)*s)) {
+            s++;
+        }
+    } while ((*d++ = *s++));
+}
+
+static void test_print_roundtrip(const char * signature) {
+    subtest(signature) {
+        plan(1);
+        infix_type * type = NULL;
+        infix_arena_t * arena = NULL;
+        infix_status status = infix_type_from_signature(&type, &arena, signature);
+
+        if (status != INFIX_SUCCESS) {
+            fail("Parsing failed, cannot test printing.");
+            infix_arena_destroy(arena);
+            return;
+        }
+
+        char buffer[1024];
+        status = infix_type_print(buffer, sizeof(buffer), type, INFIX_DIALECT_SIGNATURE);
+        if (status != INFIX_SUCCESS)
+            fail("Printing failed.");
+        else {
+            char original_normalized[1024];
+            char printed_normalized[1024];
+            snprintf(original_normalized, sizeof(original_normalized), "%s", signature);
+            snprintf(printed_normalized, sizeof(printed_normalized), "%s", buffer);
+
+            normalize_string(original_normalized);
+            normalize_string(printed_normalized);
+
+            ok(strcmp(original_normalized, printed_normalized) == 0, "Printed string should match original signature");
+            if (strcmp(original_normalized, printed_normalized) != 0) {
+                diag("Original (normalized): %s", original_normalized);
+                diag("Printed  (normalized): %s", printed_normalized);
+            }
+        }
+
+        infix_arena_destroy(arena);
+    }
+}
+
 TEST {
-    plan(5);
+    plan(6);
 
     subtest("Valid Single Types") {
         plan(15);
@@ -293,5 +343,18 @@ TEST {
                 skip(7, "Skipping detail checks");
             infix_arena_destroy(arena);
         }
+    }
+    subtest("Round trip") {
+        plan(10);
+        test_print_roundtrip("int");
+        test_print_roundtrip("*[10:{int,float}]");
+        test_print_roundtrip("<*void, double>");
+        test_print_roundtrip("(*char;int,double)->void");
+        test_print_roundtrip("{<int,char>, *char}");
+        test_print_roundtrip("struct<Node>{int,*struct<Node>}");
+        test_print_roundtrip("e:longlong");
+        test_print_roundtrip("v[4:float]");
+        test_print_roundtrip("struct<MyStruct>");
+        test_print_roundtrip("union<MyUnion>");
     }
 }
